+++ src/program/commands/asgi.rs
@@ -63,7 +63,8 @@ impl ASGIServerCommand {
 
 impl RunnableCommand for ASGIServerCommand {
     fn cli_parser(&self) -> Command {
-        HttpServerConfig::add_command_options(Command::new("runserver"))
+        HttpServerConfig::add_command_options(Command::new("serve"))
+            .about(format!("Run {} as an ASGI server.", self.app_path))
     }
 
     fn make_runnable(&mut self, args: &ArgMatches, context: PuffContext) -> Result<Runnable> {