+++ src/python/mod.rs
@@ -214,7 +214,7 @@ pub(crate) fn bootstrap_puff_globals(config: RuntimeConfig) -> PuffResult<()> {
             let locals = [("abs_file", format!("{}/bin/activate", v))].into_py_dict(py);
             py.run(ACTIVATE_THIS, None, Some(locals))?;
         }
-        
+
         let sys_path = py.import("sys")?.getattr("path")?;
         let mut paths = config.python_paths();
         paths.reverse();